csrf removed for now, structured layout tho
