+++ calendify/src/presentationLayer/views/index.hbs
@@ -129,7 +129,7 @@
       </div>
 
       <div class="column is-6">
-        <form action="" method="POST">
+        <form action="/account/login" method="POST">
           <div class="field">
             <label class="label">Email</label>
             <div class="control">
